optional compile & run TEE customized codes

+++ python/ray/_private/node.py
@@ -26,7 +26,7 @@ import ray._private.utils
 from ray._private import storage
 from ray._private.gcs_utils import GcsClient
 from ray._private.resource_spec import ResourceSpec
-from ray._private.utils import open_log, try_to_create_directory, try_to_symlink
+from ray._private.utils import open_log, ray_in_tee, try_to_create_directory, try_to_symlink
 
 # Logger for this module. It should be configured at the entry point
 # into the program using Ray. Ray configures it by default automatically
@@ -144,7 +144,7 @@ class Node:
         )
 
         self._resource_spec = None
-        self._localhost = "127.0.0.1"
+        self._localhost = socket.gethostbyname("localhost") if not ray_in_tee() else "127.0.0.1"
         self._ray_params = ray_params
         self._config = ray_params._system_config or {}
 
@@ -1061,30 +1061,36 @@ class Node:
         assert self._gcs_client is not None
         self._write_cluster_info_to_kv()
 
-        # Workaround because these params are not available in ray.init().
-        # Should be reverted once we start Ray with `ray start` CLI.
+		if not ray_in_tee():
+        	# Workaround because these params are not available in ray.init().
+        	# Should be reverted once we start Ray with `ray start` CLI.
 
-        # if not self._ray_params.no_monitor:
-        #     self.start_monitor()
+        	if not self._ray_params.no_monitor:
+        	    self.start_monitor()
 
-        # if self._ray_params.ray_client_server_port:
-        #     self.start_ray_client_server()
+        	if self._ray_params.ray_client_server_port:
+        	    self.start_ray_client_server()
 
-        # if self._ray_params.include_dashboard is None:
-        #     # Default
-        #     include_dashboard = True
-        #     raise_on_api_server_failure = False
-        # elif self._ray_params.include_dashboard is False:
-        #     include_dashboard = False
-        #     raise_on_api_server_failure = False
-        # else:
-        #     include_dashboard = True
-        #     raise_on_api_server_failure = True
+	        if self._ray_params.include_dashboard is None:
+	            # Default
+	            include_dashboard = True
+	            raise_on_api_server_failure = False
+	        elif self._ray_params.include_dashboard is False:
+	            include_dashboard = False
+	            raise_on_api_server_failure = False
+	        else:
+	            include_dashboard = True
+	            raise_on_api_server_failure = True
 
-        self.start_api_server(
-            include_dashboard=False,
-            raise_on_failure=False,
-        )
+	        self.start_api_server(
+	            include_dashboard=include_dashboard,
+	            raise_on_failure=raise_on_api_server_failure,
+	        )
+		else:
+			self.start_api_server(
+	            include_dashboard=False,
+	            raise_on_failure=False,
+	        )
 
     def start_ray_processes(self):
         """Start all of the processes on the node."""
@@ -1124,10 +1130,11 @@ class Node:
             huge_pages=self._ray_params.huge_pages,
         )
         self.start_raylet(plasma_directory, object_store_memory)
-        # Workaround because these params are not available in ray.init().
-        # Should be reverted once we start Ray with `ray start` CLI.
-        # if self._ray_params.include_log_monitor:
-        #     self.start_log_monitor()
+        if not ray_in_tee():
+            # Workaround because these params are not available in ray.init().
+            # Should be reverted once we start Ray with `ray start` CLI.
+            if self._ray_params.include_log_monitor:
+                self.start_log_monitor()
 
     def _kill_process_type(
         self,
@@ -1182,9 +1189,10 @@ class Node:
         for process_info in process_infos:
             process = process_info.process
 
-            # TODO: Use SIGTERM first if allow_graceful is True
-            os.kill(process, signal.SIGKILL)
-            return
+            if ray_in_tee():
+                # TODO: Use SIGTERM first if allow_graceful is True
+                os.kill(process, signal.SIGKILL)
+                return
 
             # Handle the case where the process has already exited.
             if process.poll() is not None:
